requirements.txt
Django==2.0.1
django-haystack==2.6.1
django-js-asset==0.1.1
django-mptt==0.9.0
django-simditor==0.0.6
easy-thumbnails==2.5
jieba==0.39
Markdown==2.6.10
mysqlclient==1.3.12
Pillow==5.0.0
pytz==2017.3
Whoosh==2.7.4
django-allauth==0.35.0
